+++ .github/release-notes/release-notes.md
@@ -4,6 +4,7 @@ Provides data factories and a test suite base class to inherit from.
 
 - [#8](https://github.com/zhofre/test-fabric/issues/8) Added data ranges and pickers.
 - [#8](https://github.com/zhofre/test-fabric/issues/8) Added data builders.
+- [#8](https://github.com/zhofre/test-fabric/issues/8) Added random number factories.
 
 ## 📦 NuGet